requirements.txt
django==1.7
django-guardian
django-mptt
six
markdown
django-extensions
django-polymorphic
django-compressor
lxml
django-bower
djangorestframework
django-filter
django-activity-stream